+++ razzle.config.js
@@ -33,7 +33,6 @@ module.exports = {
       options: {
         importLoaders: 2,
         sourceMap: true,
-        localIdentName: '[name]__[local]___[hash:base64:5]',
       },
     };
     const POST_CSS_LOADER = {
@@ -82,8 +81,6 @@ module.exports = {
               options: {
                 importLoaders: 2,
                 sourceMap: true,
-                modules: false,
-                localIdentName: '[name]__[local]___[hash:base64:5]',
               },
             },
             POST_CSS_LOADER,