+++ src/components/manage/AnchorPlugin/index.jsx
@@ -4,7 +4,6 @@ import { EditorState, Modifier } from 'draft-js';
 import DefaultLink from './components/Link';
 import LinkButton from './components/LinkButton';
 import linkStrategy, { matchesEntityType } from './linkStrategy';
-import linkStyles from './linkStyles.module.css';
 
 function removeEntity(editorState) {
   const contentState = editorState.getCurrentContent();
@@ -46,7 +45,8 @@ function removeEntity(editorState) {
 }
 
 export default (config = {}) => {
-  const defaultTheme = linkStyles;
+  // ToDo: Get rif of the remainings of having the original CSS modules
+  const defaultTheme = {};
 
   const { theme = defaultTheme, placeholder, Link, linkTarget } = config;
 
@@ -58,7 +58,7 @@ export default (config = {}) => {
         component:
           Link ||
           decorateComponentWithProps(DefaultLink, {
-            className: theme.link,
+            className: 'link-anchorlink-theme',
             target: linkTarget,
           }),
       },